+++ docs/CHANGELOG.md
@@ -18,6 +18,7 @@ Entries inside each section should be ordered by type:
 ## Python API
 
 * [Changed] More informative error messages when package construction fails ([#4216](https://github.com/quiltdata/quilt/pull/4216))
+* [Fixed] Allow S3 paths starting with `/` in `Package.set_dir()` and `Package.set()` ([#4207](https://github.com/quiltdata/quilt/pull/4207))
 
 # 6.1.0 - 2024-10-14